pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.poetry]
name = "poc-stream-processor-faust-avro"
version = "0.1.0"
description = "A Proof of Concept of a streaming application based on Faust that reads messages from a Kafka topic using Avro as deserializer."
authors = ["José Antonio Perdiguero López <perdy@perdy.io>"]

[tool.poetry.dependencies]
python = "^3.9.0"
clinner = "^1.12"
faust-streaming = "^0.4"
python-schema-registry-client = {version = "^1.8", extras = ["faust"]}
hvac = "^0.10"
backoff = "^1.10"
dataclasses-avroschema = "^0.20.2"

[tool.poetry.dev-dependencies]
faust-streaming = {version = "^0.4", extras = ["debug"]}
pytest = "^5.0"
flake8 = "^3.5"
isort = "^4.3"
pytest-xdist = "^1.22"
pytest-cov = "^2.5"
freezegun = "^0.3.15"
coverage = "^4.5"
ipython = "^7.6"
ipdb = "^0.12"
pytest-asyncio = "^0.10"
black = {version = "^19.3b0", allow-prereleases = true}
pytest-bdd = "^3.2.1"

[tool.black]
line-length = 120
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.tox
  | build
  | dist
)/
'''
